Update README.md
Jua-Bei
Description
In order to make the best decision, it’s good to have all the facts at your fingertips. We are one of the world’s best price and product comparison services, with one single goal – to guide consumers to smarter buying decisions.

Got your online shopping experience sorted.
We save your time when you when you want to get on with online shopping.It's easy, quick, offers plenty of information, and the buying process is a smooth one too.
Authors information
Project-Setup-Instructions and requirements single page application. Access of data from a public API Use of react components Incorporation of client-side routes. Initializing of Git ,git add files,git commit and git push. Deploying of the project.

Technology-Used
HTML CSS Reactjs Known-Bugs The project works perfectly.

License
Copyright © Jua Bei Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated documentation files (the "Software"), to deal in the Software without restriction, including without limitation the rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the Software, and to permit persons to whom the Software is furnished to do so, subject to the following conditions: The above copyright notice and this permission notice shall be included in all copies or substantial portions of the Software. THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.

How it works.
1.Bundle install.
2. npm start.